co2send/requirements.txt
flask
prometheus_client
pyserial
